previous_assignments/requirements.txt
pandas>=1.5.0
psycopg2-binary>=2.9.0
tabulate>=0.9.0
matplotlib>=3.8.0,<3.9.0
plotly>=5.17.0
sqlalchemy>=2.0.0
openpyxl>=3.1.0
numpy>=1.24.0,<2.0.0